Version1

+++ src/main/java/ni/org/ics/zpo/domain/Zpo02BiospecimenCollection.java
@@ -24,7 +24,7 @@ public class Zpo02BiospecimenCollection extends BaseMetaData implements Auditabl
     private String bscMatBldCol;
     private String bscMatBldTyp1;
     private String bscMatBldId1;
-    private Integer bscMatBldVol1;
+    private Double bscMatBldVol1;
     private String bscMatBldTime;
     private String bscMatBldCom;
     private String bscPerson1;
@@ -100,11 +100,11 @@ public class Zpo02BiospecimenCollection extends BaseMetaData implements Auditabl
     }
 
     @Column(name = "bsc_mat_bld_vol_1", nullable = true)
-    public Integer getBscMatBldVol1() {
+    public Double getBscMatBldVol1() {
         return bscMatBldVol1;
     }
 
-    public void setBscMatBldVol1(Integer bscMatBldVol1) {
+    public void setBscMatBldVol1(Double bscMatBldVol1) {
         this.bscMatBldVol1 = bscMatBldVol1;
     }
